pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Danielskry"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.1"
gunicorn = "^22.0.0"
fastapi = "^0.111.0"
pydantic = "^2.7.3"
langchain-community = "^0.2.4"
langchain = "^0.2.3"
openai = "^1.33.0"
langchain-openai = "^0.1.8"
langchain-core = "^0.2.5"
uvicorn = "^0.30.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
